I added the updated Class Diagram and changed the floats to decimals.

+++ src/backend/ClubManagement/ClubManagement.Domain/Player.cs
@@ -9,8 +9,8 @@ namespace ClubManagement.Domain
     public class Player : Member
     {
 
-        public float Height { get; set; }
-        public float Weight { get; set; }
+        public decimal Height { get; set; }
+        public decimal Weight { get; set; }
         public int PlayerNumber { get; set; }
     }
 }
